Separate out Consumable and optimize Concoction.
