+++ src/net/sourceforge/kolmafia/persistence/DebugDatabase.java
@@ -2739,7 +2739,7 @@ public class DebugDatabase {
 
   private static void checkConsumables(
       final PrintStream report,
-      final Collection<ConsumablesDatabase.Consumable> consumables,
+      final Collection<Consumable> consumables,
       final String tag) {
     if (consumables.size() == 0) {
       return;
@@ -2755,7 +2755,7 @@ public class DebugDatabase {
   }
 
   private static void checkConsumable(
-      final PrintStream report, ConsumablesDatabase.Consumable consumable) {
+      final PrintStream report, Consumable consumable) {
     int itemId = consumable.itemId;
     // It is valid for items to have no itemId: sushi, Cafe offerings, and so on
     String text = itemId == -1 ? "" : DebugDatabase.itemDescriptionText(itemId, false);
